backend/requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
dj-config-url==0.1.1
Django==5.0.4
django-cogwheels==0.3
django-extensions==3.2.3
django-filter==23.5
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.1
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
filetype==1.2.0
gunicorn==21.2.0
html5lib==1.1
idna==3.6
l18n==2021.3
laces==0.1.1
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pillow==10.3.0
pillow_heif==0.16.0
polib==1.2.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wagtail==6.1.1
wagtail-localize==1.9
wagtail-modeladmin==2.0.0
wagtailmenus==4.0
webencodings==0.5.1
Willow==1.8.0
